requirements.txt
Flask==1.1.1
python-telegram-bot==11.1.0
psycopg2-binary==2.8.4
apiai==1.2.3